create component in homePage and continue code homePage
