+++ src/components/containers/home/Home.jsx
@@ -1,100 +1,14 @@
-import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import tmdbApi from "../../../api/tmdbApi";
-import {
-  LASTEST_TRAILERS,
-  TRENDING,
-  POPULAR,
-} from "../../../constants/constants";
-import useFetch from "../../../hooks/useFetch";
+import React from "react";
 import Button from "../../ui/button/Button";
 import Collection from "../../ui/collection/Collection";
 import Input from "../../ui/input/Input";
-import Card from "../../ui/card/Card";
-import ListHorizontal from "../../ui/listHorizontal/ListHorizontal";
-import { ToastError } from "../../../utils/toast";
-import "./home.scss";
 import EditLeader from "../../ui/editLeader/EditLeader";
+import Trending from "./trending/Trending";
+import LastestTrailers from "./lastestTrailers/LastestTrailers";
+import Popular from "./popular/Popular";
+import "./home.scss";
 
 const Home = () => {
-  const navigate = useNavigate();
-  const [timeTrending, setTimeTrending] = useState(
-    TRENDING.listTimeTrending[0]
-  );
-  const [categoryTrailers, setCategoryTrailers] = useState(
-    LASTEST_TRAILERS.listCategory[0]
-  );
-  const [categoryPopular, setCategoryPopular] = useState(
-    POPULAR.listCategory[0]
-  );
-
-  const {
-    data: trendingData,
-    error: trendingError,
-    isLoadingTrending,
-    fetch: getTrending,
-  } = useFetch(tmdbApi.getTrending);
-
-  const {
-    data: topRatedData,
-    error: topRatedError,
-    isLoadingTopRated,
-    fetch: getTopRated,
-  } = useFetch(tmdbApi.getTopRated);
-
-  const {
-    data: popularData,
-    error: popularError,
-    isLoadingPopular,
-    fetch: getPopular,
-  } = useFetch(tmdbApi.getPopular);
-
-  // console.log(popularData);
-
-  // get trending
-  useEffect(() => {
-    if (timeTrending === "Today") {
-      getTrending({ time_window: "day" });
-    }
-    if (timeTrending === "This Week") {
-      getTrending({ time_window: "week" });
-    }
-  }, [timeTrending]);
-
-  // get trailers
-  useEffect(() => {
-    if (categoryTrailers === "On TV") {
-      getTopRated({ category: "tv", page: 1 });
-    }
-    if (categoryTrailers === "In Theaters") {
-      getTopRated({ category: "movie", page: 1 });
-    }
-  }, [categoryTrailers]);
-
-  // get popular
-  useEffect(() => {
-    if (categoryPopular === "On TV") {
-      getPopular({ category: "tv", page: 1 });
-    }
-    if (categoryPopular === "In Theaters") {
-      getPopular({ category: "movie", page: 1 });
-    }
-  }, [categoryPopular]);
-
-  const handleClickCardTrending = (item) => {
-    if (item.media_type === "movie") {
-      navigate(`/movie/${item.id}`);
-    } else {
-      ToastError("Khong code page TV Detail !!!");
-    }
-  };
-
-  const handleClickCardPopular = (item) => {
-    item.release_date
-      ? navigate(`/movie/${item.id}`)
-      : ToastError("Khong code page TV Detail !!!");
-  };
-
   return (
     <div className="homepage">
       {/* Background 1 */}
@@ -169,157 +83,50 @@ const Home = () => {
       </Collection>
 
       {/* Trending */}
-      <ListHorizontal
-        titleTab={TRENDING.title}
-        listItemTab={TRENDING.listTimeTrending}
-        itemTabActive={timeTrending}
-        onClickTab={(time) => setTimeTrending(time)}
-        styleCssTabPrimary={{
-          backgroundImage:
-            "https://www.themoviedb.org/assets/2/v4/misc/trending-bg-39afc2a5f77e31d469b25c187814c0a2efef225494c038098d62317d923f8415.svg",
-          backgroundSize: "var(--maxPrimaryPageWidth)",
-          backgroundPosition: "50% 200px",
-          colorTitle: "#000",
-          borderMain: "1px solid rgba(var(--tmdbDarkBlue), 1)",
-          textColor: "rgba(var(--tmdbDarkBlue), 1)",
-          backgroundColor: "transparent",
-          textColorActive: "#a6efc4",
-          backgroundColorActive: "rgba(var(--tmdbDarkBlue), 1)",
-        }}
-      >
-        {trendingData ? (
-          trendingData.results.map((item) => {
-            return (
-              <div className="item-trending" key={item.id}>
-                <Card
-                  width={150}
-                  heightImage={225}
-                  image={`https://www.themoviedb.org/t/p/w220_and_h330_face${item.poster_path}`}
-                  name={item.name || item.title}
-                  description={item.release_date || item.first_air_date}
-                  displayIcon
-                  vote={Math.round(item.vote_average * 10)}
-                  styleCss={{
-                    textAlign: "left",
-                    paddingContent: "24px 10px 0px 10px",
-                    fontSizeName: "16px",
-                    fontWeightName: "700",
-                    textColorName: "#000",
-                    fontSizeDesc: "16px",
-                    fontWeightDesc: "400",
-                    textColorDesc: "rgba(0,0,0,0.6)",
-                  }}
-                  displayIconCirclePercent
-                  onClickCard={() => handleClickCardTrending(item)}
-                />
-              </div>
-            );
-          })
-        ) : (
-          <div>Chua co du lieu</div>
-        )}
-      </ListHorizontal>
+      <Trending />
 
       {/* Latest Trailers */}
-      <ListHorizontal
-        titleTab={LASTEST_TRAILERS.title}
-        listItemTab={LASTEST_TRAILERS.listCategory}
-        itemTabActive={categoryTrailers}
-        onClickTab={(category) => setCategoryTrailers(category)}
-        styleCssTabPrimary={{
-          backgroundImage:
-            "https://www.themoviedb.org/t/p/w1920_and_h427_multi_faces/hUtrCoGrlo3o8juAI1ZzA7F1LX9.jpg",
-          backgroundSize: "1700px 378px",
-          backgroundPosition: "50% 0",
-          colorTitle: "#fff",
-          borderMain: "1px solid rgba(var(--tmdbLightGreen), 1)",
-          textColor: "#fff",
-          backgroundColor: "transparent",
-          textColorActive: "rgba(var(--tmdbDarkBlue), 1)",
-          backgroundColorActive:
-            "linear-gradient(to right, rgba(var(--tmdbLighterGreen), 1) 0%, rgba(var(--tmdbLightGreen), 1) 100%)",
-        }}
-      >
-        {topRatedData ? (
-          topRatedData.results.map((item) => {
-            return (
-              <div className="item-trending" key={item.id}>
-                <Card
-                  width={300}
-                  heightImage={168}
-                  image={`https://image.tmdb.org/t/p/w355_and_h200_multi_faces${item.poster_path}`}
-                  name={item.name || item.title}
-                  description="Official Trailer"
-                  styleCss={{
-                    textAlign: "center",
-                    paddingContent: "24px 10px 0px 10px",
-                    fontSizeName: "20px",
-                    fontWeightName: "600",
-                    textColorName: "#fff",
-                    fontSizeDesc: "16px",
-                    fontWeightDesc: "400",
-                    textColorDesc: "#fff",
-                  }}
-                  displayIconPlay
-                  // onClickCard={() => handleClickCardPoster(item)}
-                />
-              </div>
-            );
-          })
-        ) : (
-          <div>Chua co du lieu</div>
-        )}
-      </ListHorizontal>
+      <LastestTrailers />
 
       {/* What's Popular */}
-      <ListHorizontal
-        titleTab={POPULAR.title}
-        listItemTab={POPULAR.listCategory}
-        itemTabActive={categoryPopular}
-        onClickTab={(category) => setCategoryPopular(category)}
-        styleCssTabPrimary={{
-          colorTitle: "#000",
-          borderMain: "1px solid rgba(var(--tmdbDarkBlue), 1)",
-          textColor: "rgba(var(--tmdbDarkBlue), 1)",
-          backgroundColor: "transparent",
-          textColorActive: "#a6efc4",
-          backgroundColorActive: "rgba(var(--tmdbDarkBlue), 1)",
-        }}
-      >
-        {popularData ? (
-          popularData.results.map((item) => {
-            return (
-              <div className="item-trending" key={item.id}>
-                <Card
-                  width={150}
-                  heightImage={225}
-                  image={`https://www.themoviedb.org/t/p/w220_and_h330_face${item.poster_path}`}
-                  name={item.name || item.title}
-                  description={item.release_date || item.first_air_date}
-                  displayIcon
-                  vote={Math.round(item.vote_average * 10)}
-                  styleCss={{
-                    textAlign: "left",
-                    paddingContent: "24px 10px 0px 10px",
-                    fontSizeName: "16px",
-                    fontWeightName: "700",
-                    textColorName: "#000",
-                    fontSizeDesc: "16px",
-                    fontWeightDesc: "400",
-                    textColorDesc: "rgba(0,0,0,0.6)",
-                  }}
-                  displayIconCirclePercent
-                  onClickCard={() => handleClickCardPopular(item)}
-                />
-              </div>
-            );
-          })
-        ) : (
-          <div>Chua co du lieu</div>
-        )}
-      </ListHorizontal>
+      <Popular />
 
-      <EditLeader title="Tung" imgAvatar={"https://www.gravatar.com/avatar/3af6511cf44a709e6ae5b612903c846c.jpg?s=64"}/>
+      <div className="item_leaderboard">
+        <EditLeader
+          title="Tung"
+          imgAvatar={
+            "https://www.gravatar.com/avatar/3af6511cf44a709e6ae5b612903c846c.jpg?s=64"
+          }
+          styleCss={{
+            widthDescFirst: "200px",
+            backgroundDescFirst:
+              "linear-gradient(to right, rgba(var(--tmdbLighterGreen), 1) 0%, rgba(var(--tmdbLightGreen), 1) 100%)",
+            widthDescSecond: "390px",
+            backgroundDescSecond:
+              "linear-gradient(to right, rgba(var(--tmdbLogoOrange), 1) 0%, rgba(var(--tmdbLogoRed), 1) 100%)",
+          }}
+        />
+        <EditLeader
+          title="Tung"
+          // imgAvatar={
+          //   "https://www.gravatar.com/avatar/3af6511cf44a709e6ae5b612903c846c.jpg?s=64"
+          // }
+          styleCss={{
+            widthDescFirst: "200px",
+            backgroundDescFirst:
+              "linear-gradient(to right, rgba(var(--tmdbLighterGreen), 1) 0%, rgba(var(--tmdbLightGreen), 1) 100%)",
+            widthDescSecond: "390px",
+            backgroundDescSecond:
+              "linear-gradient(to right, rgba(var(--tmdbLogoOrange), 1) 0%, rgba(var(--tmdbLogoRed), 1) 100%)",
+          }}
+        />
+        <EditLeader
+          title="Tung"
+          imgAvatar={
+            "https://www.gravatar.com/avatar/3af6511cf44a709e6ae5b612903c846c.jpg?s=64"
+          }
+        />
+      </div>
     </div>
   );
 };
