+++ src/components/containers/home/Home.scss
@@ -9,7 +9,13 @@
     margin-left: 6px;
   }
 
-  .item-trending + .item-trending {
+  .item-card + .item-card {
     margin-left: 20px;
+  }
+
+  .item_leaderboard {
+    display: flex;
+    align-items: center;
+    flex-wrap: wrap;
   }
 }
