+++ src/components/ui/editLeader/EditLeader.jsx
@@ -2,25 +2,23 @@ import React from "react";
 import Meter from "../meter/Meter";
 import "./editLeader.scss";
 
-const EditLeader = ({ title, imgAvatar }) => {
+const EditLeader = ({ title, imgAvatar, styleCss = {} }) => {
   return (
     <div className="editLeader">
-      <div className="editLeader__avatar">
-        <img src={imgAvatar} alt={title} />
+      <div className={(imgAvatar ? "editLeader__avatarImg" : "editLeader__avatarText")}>
+        {imgAvatar ? <img src={imgAvatar} alt={title} /> : <span>{title ? title.slice(0,1).toUpperCase(): "?"}</span>}
       </div>
       <div className="editLeader__info">
-        <div className="editLeader__title">RuiZafon</div>
+        <div className="editLeader__title">{title}</div>
         <div className="edit__desc">
-         
-            <Meter
-              meterCount={273293}
-              styleCss={{ widthGauge: "200px", backgroundGauge: "red" }}
-            />
-            <Meter
-              meterCount={27323}
-              styleCss={{ widthGauge: "300px", backgroundGauge: "blue" }}
-            />
-          
+          <Meter
+            meterCount={273293}
+            styleCss={{ widthGauge: styleCss.widthDescFirst, backgroundGauge: styleCss.backgroundDescFirst }}
+          />
+          <Meter
+            meterCount={27323}
+            styleCss={{ widthGauge: styleCss.widthDescSecond, backgroundGauge: styleCss.backgroundDescSecond }}
+          />
         </div>
       </div>
     </div>
