+++ src/components/ui/editLeader/editLeader.scss
@@ -1,8 +1,12 @@
+@use "../../../scss/" as *;
+
 .editLeader {
   display: flex;
   align-items: center;
+  width: 50%;
+  margin-bottom: 20px;
 
-  .editLeader__avatar {
+  .editLeader__avatarImg, .editLeader__avatarText {
     width: 56px;
     height: 56px;
     display: flex;
@@ -14,6 +18,13 @@
       display: block;
       border-radius: 50%;
     }
+  }
+
+  .editLeader__avatarText {
+    color: $body-color;
+    background: rgba(var(--accountYellow), 1);
+    font-size: 20px;
+    border-radius: 50%;
   }
 
   .editLeader__info {
