+++ src/constants/constants.js
@@ -1,14 +1,14 @@
 export const TRENDING = {
-    title: "Trending",
-    listTimeTrending: ["Today", "This Week"]
-}
+  title: "Trending",
+  listTimeTrending: ["Today", "This Week"],
+};
 
 export const LASTEST_TRAILERS = {
-    title: "Lastest Trailers",
-    listCategory: ["On TV", "In Theaters"]
-}
+  title: "Lastest Trailers",
+  listCategory: ["On TV", "In Theaters"],
+};
 
 export const POPULAR = {
-    title: "What's Popular",
-    listCategory: ["On TV", "In Theaters"]
-}
+  title: "What's Popular",
+  listCategory: ["On TV", "In Theaters"],
+};
